// src/app/route-paths.enum.ts
export enum RoutePaths {
  Conferences = 'conferences',
  EditProject = 'projects/:id',
  NonWorkingHolidays = 'non-working-holidays',
  Salaries = 'salaries',
  EquipmentRequest = 'equipment-request',
  EquipmentCheckoutRequest = 'equipment-checkout-request',
  Equipments = 'equipments',
  Books = 'books',
  Login = 'login',
  Settings = 'settings',
  Profile = 'profile',
  Skills = 'skills',
  GoogleAuthCallback = 'connect/google/redirect',
  VacationForm = 'vacation-form',
  VacationDays = 'vacation-days',
  VacationCalendar = 'vacation-calendar',
  
}